+++ src/ExAs/Assertions/MemberAssertions/Strings/IsEmptyAssertion.cs
@@ -2,6 +2,7 @@
 using ExAs.Results;
 using ExAs.Utils;
 using ExAs.Utils.StringExtensions;
+using static ExAs.Utils.StringExtensions.StringFormattingFunctions;
 
 namespace ExAs.Assertions.MemberAssertions.Strings
 {
@@ -9,7 +10,9 @@ namespace ExAs.Assertions.MemberAssertions.Strings
     {
         public ValueAssertionResult AssertValue(string actual)
         {
-            return new ValueAssertionResult(IsEmpty(actual), actual.ToValueString(), ComposeLog.Expected("empty string"));
+            return ValueAssertionResult.Create(
+                IsEmpty(actual),
+                HarmonizeLineCount(actual.ToValueString(), ComposeLog.Expected("empty string")));
         }
 
         private static bool IsEmpty(string actual)