+++ src/ExAs/Results/ValueAssertionResult.cs
@@ -1,7 +1,14 @@
-﻿namespace ExAs.Results
+﻿using System;
+
+namespace ExAs.Results
 {
     public class ValueAssertionResult
     {
+        public static ValueAssertionResult Create(bool succeeded, Tuple<string, string> strings)
+        {
+            return new ValueAssertionResult(succeeded, strings.Item1, strings.Item2);
+        }
+
         public readonly bool succeeded;
         public readonly string actualValueString;
         public readonly string expectationString;
@@ -15,7 +22,7 @@
 
         public override string ToString()
         {
-            return string.Format("ValueAssertionResult: succeeded = {0}", succeeded);
+            return $"ValueAssertionResult: succeeded = {succeeded}";
         }
     }
 }